Resolve DNS-SD URIs correctly also if they are from a service from localhost
If a service is only available locally, through the loopback device
("lo", "localhost"), like for example the IPP-over-USB daemon ipp-usb
or a Printer Application which is not sharing printers to the network,
the service cannot be accessed through the network host name but only
through the host name "localhost".

To make CUPS' DNS-SD-service-name-based device URIs working also with
these services we check whether the service comes through the loopback
interface and in this case we set the host name in the resolved
standard device URI to "localhost".