scheduler: Admin task check: Support to use libsnapd-glib instead of snapctl
When CUPS is running in a Snap the scheduler checks client Snaps
requesting administrative tasks calling the external executable
"snapctl" which does not scale well on many inquiries in parallel.

Therefore we use an equivalent library function of libsnapd-glib,
snapd_client_run_snapctl2_sync() instead, at least if the
libsnapd-glib is new enough to already contain this function, which we
check during build, by the ./configure script. If the function is not
available we automatically fall back to a build using the "snapctl"
utility.