Test suite: Add regression test for cupsSNMP/IPPSupplies
lpadmin can set cupsSNMPSupplies and cupsIPPSupplies keywords in the PPD
file which was broken in the past (Apple Issue #5610).

Until we get rid of PPD files completely or if we use the options in the
no-PPD future, it would be great if we had a regression test in the test
suite.